requirements.txt
aiohttp==3.7.3
appdirs==1.4.4
async-timeout==3.0.1
attrs==20.3.0
beautifulsoup4==4.9.3
better-profanity==0.7.0
bs4==0.0.1
certifi==2020.11.8
cffi==1.14.4
chardet==3.0.4
click==7.1.2
comtypes==1.1.7
cssselect==1.1.0
cycler==0.10.0
deeppyer==1.0.1
discord.py==1.5.1
discord==1.0.1
discord-pretty-help==1.1.2
dnspython==2.0.0
fake-useragent==0.1.11
fast-youtube-search==0.0.8
future==0.18.2
fuzzywuzzy==0.18.0
google==3.0.0
Google-Search-API==1.1.14
googletrans==3.0.0
h11==0.11.0
h2==4.0.0
hpack==4.0.0
hstspreload==2020.11.21
httpcore==0.12.2
httpx==0.16.1
hyperframe==6.0.0
idna==2.10
kiwisolver==1.3.1
lxml==4.6.2
matplotlib==3.3.3
multidict==5.0.2
multitasking==0.0.9
numpy==1.19.2
opencv-python==4.4.0.46
pandas==1.1.4
parse==1.18.0
Pillow==8.0.1
pip==20.2.4
pip-tools==5.4.0
praw==7.1.0
prawcore==1.5.0
psycopg2-binary==2.8.6
pycparser==2.20
pyee==8.1.0
pymongo==3.11.1
PyNaCl==1.4.0
pyparsing==2.4.7
pypiwin32==223
pyppeteer==0.2.2
pyquery==1.4.3
python-dateutil==2.8.1
python-dotenv==0.15.0
python-Levenshtein==0.12.0
pyttsx3==2.90
pytz==2020.4
pywin32==300
PyYAML==5.3.1
requests==2.25.0
requests-html==0.10.0
rfc3986==1.4.0
selenium==3.141.0
setuptools==50.3.2
simplejson==3.17.2
six==1.15.0
sniffio==1.2.0
soupsieve==2.0.1
tqdm==4.54.0
typing-extensions==3.7.4.3
Unidecode==1.1.1
update-checker==0.18.0
urllib3==1.26.2
vcrpy==4.1.1
w3lib==1.22.0
Wand==0.6.4
wavelink==0.9.6
websocket-client==0.57.0
websockets==8.1
wheel==0.35.1
wrapt==1.12.1
yahoo-fin==0.8.6
yarl==1.6.3
yfinance==0.1.55
youtube-dl==2020.11.26
